+++ spec/unit/substation/chain/dsl/chain_spec.rb
@@ -5,7 +5,8 @@ require 'spec_helper'
 describe Chain::DSL, '#chain' do
   subject { object.chain(other) }
 
-  let(:object)    { described_class.new(chain) }
+  let(:object)    { described_class.new(env, chain) }
+  let(:env)       { Spec::FAKE_ENV }
   let(:chain)     { Chain::EMPTY }
   let(:other)     { [ processor ] }
   let(:processor) { Spec::FAKE_PROCESSOR }