+++ spec/unit/substation/chain/dsl/class_methods/processors_spec.rb
@@ -4,10 +4,11 @@ require 'spec_helper'
 
 describe Chain::DSL, '.processors' do
 
+  let(:env)   { Spec::FAKE_ENV }
   let(:chain) { Chain::EMPTY }
 
   context "and a block is given" do
-    subject { described_class.processors(chain, &block) }
+    subject { described_class.processors(env, chain, &block) }
 
     let(:block)     { lambda { |_| use(Spec::FAKE_PROCESSOR) } }
     let(:processor) { Spec::FAKE_PROCESSOR }
@@ -16,7 +17,7 @@ describe Chain::DSL, '.processors' do
   end
 
   context "and no block is given" do
-    subject { described_class.processors(chain) }
+    subject { described_class.processors(env, chain) }
 
     it { should be_empty }
   end