+++ spec/unit/substation/chain/dsl/initialize_spec.rb
@@ -2,20 +2,18 @@
 
 require 'spec_helper'
 
-describe Chain::DSL::Builder, '#dsl' do
+# Test wether env is correctly passed down
+describe Chain::DSL, '#initialize' do
   subject { dsl.new(env, processors, &block) }
 
   let(:env)        { Spec::FAKE_ENV }
   let(:dsl)        { builder.dsl }
-  let(:builder)    { described_class.new(registry) }
+  let(:builder)    { described_class::Builder.new(registry) }
   let(:registry)   { { :test => Spec::Processor } }
   let(:processors) { [] }
   let(:block)      { lambda { |_| test(Spec::FAKE_HANDLER) } }
+
   let(:processor)  { Spec::FAKE_PROCESSOR }
 
   its(:processors) { should include(processor) }
-
-  it "should create a subclass of Chain::DSL" do
-    subject.class.should be < Chain::DSL
-  end
 end