+++ spec/unit/substation/environment/chain_spec.rb
@@ -4,16 +4,19 @@ describe Substation::Environment, '#chain' do
 
   let(:object) { described_class.build(&block) }
 
-  let(:expected) { Chain.build(dsl, other, &chain) }
   let(:other)    { Chain::EMPTY }
   let(:chain)    { lambda { |_| test Spec::FAKE_HANDLER } }
   let(:dsl)      { Chain::DSL::Builder.call(registry) }
   let(:registry) { described_class::DSL.registry(&block) }
   let(:block)    { lambda { |_| register(:test, Spec::Processor) } }
 
+  let(:expected) { Chain.new(processors) }
+
   context "when other is not given" do
     context "and a block is given" do
       subject { object.chain(&chain) }
+
+      let(:processors) { dsl.processors(object, other, &chain) }
 
       it { should eql(expected) }
     end
@@ -21,7 +24,7 @@ describe Substation::Environment, '#chain' do
     context "and no block is given" do
       subject { object.chain }
 
-      let(:expected) { Chain.build(dsl, other) }
+      let(:processors) { dsl.processors(object, other) }
 
       it { should eql(expected) }
     end
@@ -31,13 +34,15 @@ describe Substation::Environment, '#chain' do
     context "and a block is given" do
       subject { object.chain(other, &chain) }
 
+      let(:processors) { dsl.processors(object, other, &chain) }
+
       it { should eql(expected) }
     end
 
     context "and no block is given" do
       subject { object.chain(other) }
 
-      let(:expected) { Chain.build(dsl, other) }
+      let(:processors) { dsl.processors(object, other) }
 
       it { should eql(expected) }
     end