+++ spec/unit/substation/processor/evaluator/call_spec.rb
@@ -5,21 +5,45 @@ require 'spec_helper'
 describe Processor::Evaluator, '#call' do
   subject { object.call(request) }
 
-  let(:object)  { described_class.new(Spec::Handler::Evaluator.new) }
-  let(:request) { Request.new(env, input) }
-  let(:env)     { mock }
+  let(:request) { Request.new(mock('env'), input) }
 
-  context "when evaluation is successful" do
-    let(:input)    { :success }
-    let(:response) { Response::Success.new(request, input) }
+  context "when no failure chain is registered" do
 
-    it { should eql(response) }
+    let(:object) { described_class.new(Spec::FAKE_ENV, Spec::Handler::Evaluator.new) }
+
+    context "when evaluation is successful" do
+      let(:input)    { :success }
+      let(:response) { Response::Success.new(request, input) }
+
+      it { should eql(response) }
+    end
+
+    context "when evaluation is not successful" do
+      let(:input)    { :invalid }
+      let(:response) { Response::Failure.new(request, :failure) }
+
+      it { should eql(response) }
+    end
   end
 
-  context "when evaluation is not successful" do
-    let(:input)    { :invalid }
-    let(:response) { Response::Failure.new(request, :failure) }
+  context "when a failure chain is registered" do
 
-    it { should eql(response) }
+    let(:object) { described_class.new(env, Spec::Handler::Evaluator.new, &block) }
+    let(:env)    { Environment.new({}) }
+    let(:block)  { lambda { |_| use(Processor::Wrapper.new(env, Spec::Presenter)) } }
+
+    context "when evaluation is successful" do
+      let(:input)    { :success }
+      let(:response) { Response::Success.new(request, input) }
+
+      it { should eql(response) }
+    end
+
+    context "when evaluation is not successful" do
+      let(:input)    { :invalid }
+      let(:response) { Response::Failure.new(request, Spec::Presenter.new(:failure)) }
+
+      it { should eql(response) }
+    end
   end
 end